feat: add select component in ui-kit and refactor organizations-relationship
