+++ packages/apps/spaces/components/ui-kit/atoms/index.ts
@@ -19,3 +19,4 @@ export { Tooltip } from './tooltip';
 export { Ribbon } from './ribbon';
 export { FileUpload } from './file-upload';
 export * from './icons';
+export { Select, useSelect } from './select';
