+++ packages/apps/spaces/components/ui-kit/atoms/select/index.ts
@@ -1,0 +1,2 @@
+export { Select } from './Select';
+export { useSelect } from './useSelect';
