+++ packages/apps/spaces/components/ui-kit/atoms/select/useSelect.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { SelectContext } from './context';
+
+export const useSelect = () => useContext(SelectContext);
